requirements.txt
fastapi==0.115.6
langchain==0.3.16
langchain-core==0.3.32
langchain-google-vertexai==2.0.9
langchain-openai==0.3.0
google-cloud-aiplatform==1.75.0
mlflow
pydantic==2.9.2
pydantic_core==2.23.4
uvicorn==0.34.0
starlette==0.41.3
anyio==4.8.0
sniffio==1.3.1
h11==0.14.0
typing_extensions==4.12.2
annotated-types==0.7.0
jsonpatch==1.33
jsonpointer==3.0.0
python-multipart==0.0.20
click==8.1.8
colorama==0.4.6
httpx==0.27.2
httpcore==1.0.7
certifi==2024.12.14
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.10
requests==2.32.3
urllib3==2.3.0
charset-normalizer==3.4.1
openai==1.59.7
tiktoken==0.8.0
regex==2024.11.6
distro==1.9.0
jiter==0.8.2
google-auth==2.37.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
rsa==4.9
cachetools==5.5.0
google-api-core==2.24.0
googleapis-common-protos==1.66.0
proto-plus==1.25.0
protobuf==5.29.2
grpcio==1.69.0
grpcio-status==1.69.0
packaging==24.2
python-dateutil==2.9.0.post0
six==1.17.0
pytz==2024.2
numpy==1.26.4
pandas==2.2.3
tzdata==2024.2
scipy==1.15.1
tenacity==9.0.0
langsmith==0.2.10
orjson==3.10.13
requests-toolbelt==1.0.0
google-cloud-core==2.4.1
google-resumable-media==2.7.2
google-crc32c==1.6.0
google-cloud-storage==2.19.0
shapely==2.0.6
PyYAML==6.0.2
sqlalchemy==2.0.36
greenlet==3.1.1
alembic==1.13.3
mako==1.3.5
flask==3.0.3
werkzeug==3.0.4
itsdangerous==2.2.0
blinker==1.8.2
jinja2==3.1.4
markupsafe==2.1.5
docker==7.1.0
entrypoints==0.4
gitpython==3.1.43
gitdb==4.0.11
smmap==5.0.1
pyarrow==17.0.0
querystring-parser==1.2.4
databricks-sdk==0.38.0
cloudpickle==3.1.0
pymongo
google-genai
